+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fixed bug in `search_environment_data()` where evaluating some objects that are not data frames resulted in an error.
 * Updated `search_environment_data()` so each search type is optional.
+* Updated `cnt_search_result()` to optionally ignore results by name.
 
 # dmcognigen 0.1.1